+++ R/data.R
@@ -1,8 +1,8 @@
 #' Valid reconstructable feature collections of the GPlates Web Service
 #'
-#' The object contains valid returns as of 2023-07-13.
+#' The object contains valid returns as of 2024-02-02.
 #'
-#' The valid return combinations for selected models and feature collections are copied from \url{https://gwsdoc.gplates.org/reconstruction-models}. 
+#' The valid return combinations for selected models and feature collections are copied from \url{https://gwsdoc.gplates.org/models}. 
 #'
 #' \describe{
 #' 	\item{\code{model}}{The name of the reconstruction model.}